fixes the slice copying of group records
[body]

This and the last few commits were addressing an issue where when we
were do the run copy slice routine it was vry slow and the files were
ballooning in size.

THe problem fundamentally was that in the variable length records the
datatype was not being set correctly.

THis has been corrected and the proper (as best as possible) setting
of compression and chunking was set.