adds BaseContigTree for serializability
[body]

The idea I wanted was to be able to serialize a contigtree so I could
cache them and avoid recomputation of the tree everytime I create the
dang thing.

So I refactored out all of the tree computations and made a class that
doesn't keep the HDF5 reference, called BaseContigTree, which can be
serialized. Then I made it so that the ContigTree can just be
constructed from a BaseContigTree and the original WepyHDF5.

Also added a context manager to the ContigTree which helps a lot with
analysis.