adds option to force setting of traj field data
THis is only available from the internal hidden interface. Basically
it turns off checking for the data constraints (which is very weak and
heuristic to begin with and doesn't handle trajectories of different
lengths).

This was motivated by wanting to set dirty data, which I knew was okay
and just wanted to turn that off.

Users will not notice unless they dig