refactors and implements task work mapper cleanly
Reworked a lot of the code related to the work mapper. Factored out
the runner specific stuff into task process class and made subclasses
for OpenMM that are specific to the different platforms with their
options.

The code is actually exactly the same as the other worker processes,
interestingly.

A lot of the stuff from before wasn't really necessary like locking
with semaphores and using a queue at the same time.

I also clarified my understanding and removed the stuff related to the
serialization stuff.

There shouldn't be a need to explicitly unwrap walker states and use
the underlyning __setstate__ and other methods because these are
exactly what pickle is using. If we do this then the unpacking of
results needs to have special handler logic that is called by the
mapper, which is undesirable. Instead we should just rely on the
pickle facilities, which basically is doing this for us.

Not sure what the test case was before that led to thinking that was
faster, but I would like to see performance changes worked on with my
draft since we can now isolate the effects.