fixes bug with deleting last checkpoint snapshot on the frequency
There was a corner case where the last snapshot would get deleted
despite the run being saved. This occured due to the way the
checkpoint orchestrator is trimmed to only include the run
involved. If the last cycle had a checkpoint then it would get the
hash of that as both the final snapshot and the n-1 snapshot. Then at
the end it would delete the n-1 snapshot.

In this we just skip the final save checkpoint step if the cycle was a
checkpoint one.