added option to set the platform at run time
first we had to add an option for "apparatus_opts" into the
configuration so that it can be injected at runtime. And also had to
handle the reparametrization for this.

Second we added a keyword argument to the sim_manager.run_cycle
command to allow passing the runner options in.

3rd we implemented setting a pre_cycle platform and kwargs which is
then resolved when runner.run_segment is run.

It is accompanied by a test.