profiling wepy simulations, add sim monitor hook to run_simulation
THe major thing that was added

too much to describe in headline exactly. Here is what I was doing.

I had prototyped the prometheus stuff in the wumpus project and moved
it over here, so there is a bunch of stuff here from that. In the
'prom' jig I am working on getting all kinds of different profiling
things.

First I just got the basic prometheus metrics from the node and
GPU.

Then I was experimenting with the number of cycles which was a point
of confusion which is irrelevant though.

Then I started running the filprofiler. This didn't actually help me
so much since I am not looking for memory bottlenecks so much as a
"memory leak" of sorts.

So I went to the pympler tool to try and get some basic info on the
size of some objects to see if I couldn't see if there was anything
large or growing during a simulation.

I figured the basics of this out, and then I made gauges for all of
those via prometheus and made a dashboard for them in grafana.

From this I can find that definitely there is a "leak" in the
dashboard reporter. There are other things I could monitor, but I kind
of just want to stick to the basics.