weighted transition counts, network handles assignments flexibly
Three major things:

- new weights option for the transition counts function
- MacroStateNetwork calculates counts as weighted counts now
- MacroStateNetwork handles assignment fields more flexibly

For the first one you can now specify a big matrix the same shape as
the assignments that gives the weights of all the walkers. The weight
instead of being a count of 1.0 will be the weight of the walker in
the start position of the transition.

The macrostate network now uses this weighted counts as the way it
generates the counts matrix.

Lastly, it was always a PITA with regards to the shape of the field
datum for the assignments. In my examples (and I suspect in many other
places) it used unwrapped (e.g. (10,) arrays) but really it should
have used wrapped (e.g. (10, 1) arrays) as per my requests in the wepy code. But since there
is no typing... it fell off the rails. So I decided to make it accept
both.