Fixes ContigTreeProfiler.fe_profile_all
This was doing the wrong thing in the past.

I rewrote it from using the trace to just getting all the data as
trajectory sized chunks, I think this will be faster and more memory
aligned.

Updated the changelog.

Also cleaned up a few formatting things elsewhere that was related