fixes arg passing in MSN.node_fields_map
The docstring said that the mapped function should receive the node_id
as the first argument but the code wasn't actually doing that.

The node_id was added into that code so that now the docstring is
correct.

This makes sense, since when mapping it would be nice to know which
node you are so you can do something about that with the rest of the
network or use a lookup table in the arguments