rework transition counts and probs
This was a flawed system in the past and needed redoing.

Before the sum of the weighted counts would be made and then called
the one and only "countsmat".

This has been replaced by 3 different equally important values:

- weighted_counts :: the same thing
- unweighted_counts :: unweighted version of the same thing,
- essentially number of samples
- all_transitions :: array of all of the weights of the transitions.

The reasoning being is that later when making markov state models you
would want to choose between these for different processes.

For instance a weighted count with only one sample may want to be
trimmed off before doing markov state modelling because there is little
confidence in one sample

That is also the motivation behind providing the whole sample set of
transitions. Do stats on it if you want, averages, variances, etc.

This also deprecates the interface for the 'probmat' and 'countsmat'.