adds preliminary support for variable number of walkers in WepyHDF5
This is a WIP and subject to change.

WARNING: This does not yet make all of the APIs aware of variable
numbers of walkers. There are likely bugs in the analysis code if you
have variable number of walkers. This support will likely get added
after this is merged to master.

What has changed?

First in the WepyHDF5 class itself:

- There are a number of new methods to support proper and robust frame/cycles:
  - run_cycle_idxs
  - run_cycle_walker_idxs
  - run_cycle_walker_traces
  - get_traj_field_names
  - get_traj_field_names_observables
  - get_traj_field_names_alt_reps
  - get_traj_cycle_idxs
  - get_traj_field_num_frames

- A number of methods were altered to not be lazy about getting
  frame/cycle idxs etc.
  - num_traj_frames
  - get_traj_field_cycle_idxs

The main method of supporting variable numbers of walkers is through
sparse fields over all of the fields.

These latter methods also now take into account the offset in the
beginning of a trajectory for a new larger number of
walkers.

This is both for contiguous and sparse fields.

This relies on the presence of a metadata attribute in each traj group
that is the 'start_cycle_idx' that is set in the reporter.

Additionally some small changes:

- add_traj now returns the traj_grp that it created

For the WepyHDF5Reporter there are some additions to the API.

Mainly this is the semantics around specifying the 'sparse_fields'
argument. It should be backwards compatible.

But now there is the option of both specifying a single Ellipsis,
which makes all fields sparse.

This would be the default/base recommendation for a variable number of
walker simulation.

And now also there are 3 total choices for a field by field specification:

- `int` for a frequency
- `None` if it is not sparse (ensures it is not sparse. Can act as an "exclude")
- `Ellipsis` which will initialize the field as sparse but will still save everything that it can.

It is the `Ellipsis` option which is used for the variable number of walkers.

It should be possible to always make this the default. But likely
there is too much code relying on the datasets being uncomplicated by
sparse fields.

Additionally the alt_reps and all_atoms_rep_freq options support the
Ellipsis option as well now.

The other major change is that now the reporter will record much more
metadata on the cycle indices for the runs and trajectories.

The new trajectory group metadata fields are:

- 'first_cycle_idx' :: previously it did set this as 'cycle_idx' kind
  of as a hail mary for supporting variable walkers.
- 'num_cycles' :: Which gets updated with each call to report
- 'last_cycle_idx' :: which also gets updated with each call to report

This allows for some consistency checks with both the shape of the
data and in a run as a whole which is much more efficient than
actually counting them.

Currently 'first_cycle_idx' is the only actualized dependency in the
WepyHDF5 class used as offsets for cycle indices.

The new run group metadata attributes are:

- 'num_cycles'
- 'last_cycle_idx'