fixes the way "frames" specs are handled, fix sparse weights
The `get_traj_field` was broken with the new cycle_idxs semantics.

First the weights were not sparse as they should always be saved.

The current commit is a HACK to get around this. There is a shim
function for getting the weights specifically as a field which will
look at the sparse idxs of the 'positions' field.

As noted in the issue this should change in the future in a more
robust mechanism.

We also have a split in the semantics of the 'frames' argument to
`get_trajectory_field`. For that function it is actually the
"cycle_idxs" as one would logically expect.

The lower level functions for contiguous and sparse fields accept a
"frames" but this is actually the 'frame_idx' which is just the
counting index of the frames in that lane.