clarifies selection arguments for `get_traj_fields`
Previously there was just a 'frames' argument for choosing a subset of
the structures in a trajectory.

With the introduction of "sparse trajectories" and not just "sparse
fields" there is now a distinction between 'frame_idxs' and
'cycle_idxs'.

This was addressed in a lower level in earlier commits but the
`get_traj_fields` API stayed the same with the 'frames' argument
simply meaning 'cycle_idxs'.

Components that were using `get_traj_fields` (like in the distributed
analysis chunker) were using 'frames' in an ambiguous way since there
functionally wasn't a difference.

Now this difference matters. For tools like the chunker where the
actual time dependence between walkers is irrelevant they needed a way
to query just on the data shape itself, or the 'frame_idxs'.

'frame_idxs' was added as an argument to support this.

The 'frames' argument is a bit redundant and confusing with this (and
not clear it means actuall cycle_idxs) so we deprecated it and
replaced it with 'cycle_idxs'. It still exists for now but raises a
deprecation warning on usage and may be removed in the future.